Add enmu Command class
The use of enum Command was troublesome when placinge it in Gloqi
class because whenever my CommandParser class needs to access
Command, it need to make a reference to Gloqi class.

Let's make enum Command into a class so that it can be access
conveniently by both class and also future class.